requirements.txt
asgiref==3.2.10
Django==3.1.1
gunicorn==20.0.4
whitenoise==5.2.0
djangorestframework==3.11.1
django-heroku==0.3.1
fcm-django==0.3.7
Pillow==5.1.0
django-cloudinary-storage==0.3.0
cloudinary==1.25.0
django-simple-history==2.12.0
